requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
